refactor: change stepCost to long for consistency

+++ app/src/main/java/io/github/sasori_256/town_planning/entity/resident/strategy/RandomMoveAction.java
@@ -232,7 +232,7 @@ public class RandomMoveAction implements GameAction {
           continue;
         }
         MapCell cell = map.getCell(new Point2D.Double(nx, ny));
-        int stepCost = cell.getMoveCost();
+        long stepCost = cell.getMoveCost();
         if (stepCost >= COST_INF) {
           continue;
         }
